coding: use ASN1_MAX_TAG_SIZE consistently
There were two different definitions of the maximum tag size in DER,
namely, MAX_TAG_LEN (= 16) and ASN1_MAX_TAG_SIZE (= 4), where all
use-cases of the former is to create a buffer for _asn1_tag_der, which
takes an argument with the latter length. This replaces the usage of
the former with the latter.

Signed-off-by: Daiki Ueno <ueno@gnu.org>