cicd: Rework workflow rules.
We want non-standard (tagged) runner jobs like macOS/Windows/OpenBSD
to run on our mainline branch and in merge requests (to access the
runners in our namespace), but not in forks which typically doesn't
have those runners available.  Achieving this while avoiding duplicate
pipelines is a bit tricky.  It is complicated that we at the same time
want the Coverity job to run on our mainline branch but it rarely make
sense to use in forks.

Signed-off-by: Simon Josefsson <simon@josefsson.org>